+++ src/lib/index.ts
@@ -6,4 +6,6 @@ import AccountingDiaryWrapper from "./AccountingDiaryWrapper.tsx";
 import { GlobalProvider } from "./context.tsx";
 
 export { Header, Footer, DialogOperation, AccountingDiary, AccountingDiaryWrapper, GlobalProvider };
+export type { IDataItem, IStyleConfig, ILabels, SortField, SortOrder } from '../types/common';
+export { defaultLabels } from '../types/common';
 export default AccountingDiaryWrapper;